Update dom manipulation
Remove start button onclick event from index.js
Remove displayBoard from game.js
Remove hidePlayerTurn from game.js
Add to doman.js module
