+++ index.html
@@ -29,7 +29,7 @@
         </form>
       </div>
       <div><p id="message"></p></div>
-      <div id="board" class="mx-auto">
+      <div id="board" class="mx-auto" hidden>
         <div class="d-flex flex-row justify-content-center">
           <button id="pos0" class="cell" disabled></button>
           <button id="pos1" class="cell" disabled></button>
