+++ js/modules/player.js
@@ -1,0 +1,3 @@
+const Player = (id, name, symbol) => ({ id, name, symbol });
+
+export default Player;
